completed step #7

+++ app/src/main/java/hexlet/code/model/UrlCheck.java
@@ -1,9 +1,10 @@
 package hexlet.code.model;
 
-import java.sql.Timestamp;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
-import lombok.AllArgsConstructor;
+
+import java.sql.Timestamp;
 
 @AllArgsConstructor
 @Getter
@@ -14,10 +15,10 @@ public final class UrlCheck {
     private String title;
     private String h1;
     private String description;
-    private int urlId;
+    private long urlId;
     private Timestamp createdAt;
 
-    public UrlCheck(int urlId, int statusCode) {
+    public UrlCheck(long urlId, int statusCode) {
         this.statusCode = statusCode;
         this.urlId = urlId;
     }
